detail 画面をモジュール化
