+++ Pokdex/.idea/gradle.xml
@@ -16,6 +16,7 @@
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/core/data" />
             <option value="$PROJECT_DIR$/feature" />
+            <option value="$PROJECT_DIR$/feature/pokemondetail" />
             <option value="$PROJECT_DIR$/feature/pokemonlist" />
           </set>
         </option>
