+++ Pokdex/app/build.gradle
@@ -100,6 +100,7 @@ dependencies {
     implementation 'androidx.hilt:hilt-navigation-compose:1.0.0-alpha03'
 
     // modules
+    implementation project(":feature:pokemondetail")
     implementation project(":feature:pokemonlist")
     implementation project(":core:data")
 }
