+++ Pokdex/app/src/main/java/com/example/pokdex/MainActivity.kt
@@ -10,7 +10,6 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.example.pokdex.pokemondetail.PokemonDetailScreen
 import jp.mydns.kokoichi0206.pokemonlist.PokemonListScreen
 import com.example.pokdex.ui.theme.JetpackComposePokedexTheme
 import dagger.hilt.android.AndroidEntryPoint
@@ -48,7 +47,7 @@ class MainActivity : ComponentActivity() {
                         val pokemonName = remember {
                             it.arguments?.getString("pokemonName")
                         }
-                        PokemonDetailScreen(
+                        jp.mydns.kokoichi0206.pokemondetail.PokemonDetailScreen(
                             dominantColor = dominantColor,
                             pokemonName = pokemonName?.toLowerCase(Locale.ROOT) ?: "",
                             navController = navController
