+++ Pokdex/feature/pokemondetail/.gitignore
@@ -1,0 +1,1 @@
+/build
