+++ Pokdex/feature/pokemondetail/src/main/java/jp/mydns/kokoichi0206/pokemondetail/ui/theme/Color.kt
@@ -1,0 +1,31 @@
+package jp.mydns.kokoichi0206.pokemondetail.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val lightGrey = Color(0xFFAAAAAA)
+
+val HPColor = Color(0xFFF5FF00)
+val AtkColor = Color(1f, 0f, 0f, 0.66f)
+val DefColor = Color(0f, 0f, 1f, 0.44f)
+val SpAtkColor = Color(0.671f, 0f, 1f, 0.57f)
+val SpDefColor = Color(1f, 0f, 0.8f, 0.7f)
+val SpdColor = Color(0f, 1f, 0.063f, 0.55f)
+
+val TypeNormal = Color(0xFFA8A77A)
+val TypeFire = Color(0xFFEE8130)
+val TypeWater = Color(0xFF6390F0)
+val TypeElectric = Color(0xFFF7D02C)
+val TypeGrass = Color(0xFF7AC74C)
+val TypeIce = Color(0xFF96D9D6)
+val TypeFighting = Color(0xFFC22E28)
+val TypePoison = Color(0xFFA33EA1)
+val TypeGround = Color(0xFFE2BF65)
+val TypeFlying = Color(0xFFA98FF3)
+val TypePsychic = Color(0xFFF95587)
+val TypeBug = Color(0xFFA6B91A)
+val TypeRock = Color(0xFFB6A136)
+val TypeGhost = Color(0xFF735797)
+val TypeDragon = Color(0xFF6F35FC)
+val TypeDark = Color(0xFF705746)
+val TypeSteel = Color(0xFFB7B7CE)
+val TypeFairy = Color(0xFFD685AD)
