+++ Pokdex/feature/pokemondetail/src/main/res/drawable/ic_weight.xml
