+++ Pokdex/settings.gradle
@@ -11,3 +11,4 @@ include ':app'
 include ':core:data'
 include ':buildsrc'
 include ':feature:pokemonlist'
+include ':feature:pokemondetail'
